Deploy backend

+++ src/main/java/com/mallang/backend/domain/Review.java
@@ -20,7 +20,7 @@ public class Review {
     private Long memberName; // 사용자 이름
     private Long doctorId; // 의사 ID
     private Long doctorName; // 의사 이름
-    private Long star; // 전체 별점
+    private int star; // 전체 별점
     private Long departmentId; // 진료과 ID
     private String departmentName; // 진료과 이름
